[Change] move extract code
