+++ docs/api/gen_doc.py
@@ -1016,173 +1016,37 @@ def filter_out_object_of_api_info_dict():
             del api_info_dict[id_api]['object']
 
 
-def extract_code_blocks_from_docstr(docstr, google_style=True):
-    """
-    extract code-blocks from the given docstring.
-    DON'T include the multiline-string definition in code-blocks.
-    The *Examples* section must be the last.
-    Args:
-        docstr(str): docstring
-        google_style(bool): if not use google_style, the code blocks will be extracted from all the parts of docstring.
-    Return:
-        code_blocks: A list of code-blocks, indent removed.
-                     element {'name': the code-block's name, 'id': sequence id.
-                              'codes': codes, 'required': 'gpu', 'in_examples': bool, code block in `Examples` or not,}
-    """
-    code_blocks = []
-
-    mo = re.search(r"Examples?:", docstr)
-
-    if google_style and mo is None:
-        return code_blocks
-
-    example_start = len(docstr) if mo is None else mo.start()
-    docstr_describe = docstr[:example_start].splitlines()
-    docstr_examples = docstr[example_start:].splitlines()
-
-    docstr_list = []
-    if google_style:
-        example_lineno = 0
-        docstr_list = docstr_examples
-    else:
-        example_lineno = len(docstr_describe)
-        docstr_list = docstr_describe + docstr_examples
-
-    lastlineindex = len(docstr_list) - 1
-
-    cb_start_pat = re.compile(r"code-block::\s*python")
-    cb_param_pat = re.compile(r"^\s*:(\w+):\s*(\S*)\s*$")
-    cb_required_pat = re.compile(r"^\s*#\s*require[s|d]\s*:\s*(\S+)\s*$")
-
-    cb_info = {}
-    cb_info['cb_started'] = False
-    cb_info['cb_cur'] = []
-    cb_info['cb_cur_indent'] = -1
-    cb_info['cb_cur_name'] = None
-    cb_info['cb_cur_seq_id'] = 0
-    cb_info['cb_required'] = None
-
-    def _cb_started():
-        # nonlocal cb_started, cb_cur_name, cb_required, cb_cur_seq_id
-        cb_info['cb_started'] = True
-        cb_info['cb_cur_seq_id'] += 1
-        cb_info['cb_cur_name'] = None
-        cb_info['cb_required'] = None
-
-    def _append_code_block(in_examples):
-        # nonlocal code_blocks, cb_cur, cb_cur_name, cb_cur_seq_id, cb_required
-        code_blocks.append(
-            {
-                'codes': inspect.cleandoc("\n" + "\n".join(cb_info['cb_cur'])),
-                'name': cb_info['cb_cur_name'],
-                'id': cb_info['cb_cur_seq_id'],
-                'required': cb_info['cb_required'],
-                'in_examples': in_examples,
-            }
-        )
-
-    for lineno, linecont in enumerate(docstr_list):
-        if re.search(cb_start_pat, linecont):
-            if not cb_info['cb_started']:
-                _cb_started()
-                continue
-            else:
-                # cur block end
-                if len(cb_info['cb_cur']):
-                    _append_code_block(lineno > example_lineno)
-                _cb_started()  # another block started
-                cb_info['cb_cur_indent'] = -1
-                cb_info['cb_cur'] = []
-        else:
-            if cb_info['cb_started']:
-                # handle the code-block directive's options
-                mo_p = cb_param_pat.match(linecont)
-                if mo_p:
-                    if mo_p.group(1) == 'name':
-                        cb_info['cb_cur_name'] = mo_p.group(2)
-                    continue
-                # read the required directive
-                mo_r = cb_required_pat.match(linecont)
-                if mo_r:
-                    cb_info['cb_required'] = mo_r.group(1)
-                # docstring end
-                if lineno == lastlineindex:
-                    mo = re.search(r"\S", linecont)
-                    if (
-                        mo is not None
-                        and cb_info['cb_cur_indent'] <= mo.start()
-                    ):
-                        cb_info['cb_cur'].append(linecont)
-                    if len(cb_info['cb_cur']):
-                        _append_code_block(lineno > example_lineno)
-                    break
-                # check indent for cur block start and end.
-                if cb_info['cb_cur_indent'] < 0:
-                    mo = re.search(r"\S", linecont)
-                    if mo is None:
-                        continue
-                    # find the first non empty line
-                    cb_info['cb_cur_indent'] = mo.start()
-                    cb_info['cb_cur'].append(linecont)
-                else:
-                    mo = re.search(r"\S", linecont)
-                    if mo is None:
-                        cb_info['cb_cur'].append(linecont)
-                        continue
-                    if cb_info['cb_cur_indent'] <= mo.start():
-                        cb_info['cb_cur'].append(linecont)
-                    else:
-                        if linecont[mo.start()] == '#':
-                            continue
-                        else:
-                            # block end
-                            if len(cb_info['cb_cur']):
-                                _append_code_block(lineno > example_lineno)
-                            cb_info['cb_started'] = False
-                            cb_info['cb_cur_indent'] = -1
-                            cb_info['cb_cur'] = []
-    return code_blocks
-
-
 def reset_api_info_dict():
     global api_info_dict, parsed_mods
     api_info_dict = {}
     parsed_mods = {}
 
 
-arguments = [
-    # flags, dest, type, default, help
-    ['--logf', 'logf', str, None, 'file for logging'],
-    [
-        '--attr',
-        'travelled_attr',
-        str,
-        'all,dict',
-        'the attribute for travelling, must be subset of [all,dict], such as "all" or "dict" or "all,dict".',
-    ],
-    [
-        '--gen-rst',
-        'gen_rst',
-        bool,
-        True,
-        'generate English api reST files. If "all" in attr, only for "all".',
-    ],
-]
-
-
 def parse_args():
     """
     Parse input arguments
     """
-    global arguments
     parser = argparse.ArgumentParser(
         description='generate the api_info json and generate the English api_doc reST files.'
     )
     parser.add_argument('--debug', dest='debug', action="store_true")
-    for item in arguments:
-        parser.add_argument(
-            item[0], dest=item[1], help=item[4], type=item[2], default=item[3]
-        )
+    parser.add_argument(
+        '--logf', dest='logf', type=str, default=None, help='file for logging'
+    )
+    parser.add_argument(
+        '--attr',
+        dest='travelled_attr',
+        type=str,
+        default='all,dict',
+        help='the attribute for travelling, must be subset of [all,dict], such as "all" or "dict" or "all,dict".',
+    )
+    parser.add_argument(
+        '--gen-rst',
+        dest='gen_rst',
+        type=bool,
+        default=True,
+        help='generate English api reST files. If "all" in attr, only for "all".',
+    )
 
     args = parser.parse_args()
     return args
